+++ src/controllers/index.ts
@@ -1,0 +1,2 @@
+export * from './add-client.ctrl';
+export * from './validate-command.ctrl';
